% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{absolutePanel}
\alias{absolutePanel}
\title{Modified absolutPanel from shiny}
\usage{
absolutePanel(
  ...,
  top = NULL,
  left = NULL,
  right = NULL,
  bottom = NULL,
  width = NULL,
  height = NULL,
  draggable = FALSE,
  fixed = FALSE,
  cursor = c("auto", "move", "default", "inherit")
)
}
\description{
Modified absolutPanel from shiny
}
