% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{fileInput}
\alias{fileInput}
\title{Modified shiny fileInput func}
\usage{
fileInput(
  inputId,
  label,
  multiple = FALSE,
  accept = NULL,
  width = NULL,
  buttonLabel = "Browse...",
  placeholder = "No file selected",
  tooltip = T,
  title = "?",
  popup = "Help tips",
  pos = "right"
)
}
\description{
Modified shiny fileInput func
}
