% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{modalDialog}
\alias{modalDialog}
\title{Mopdified modalDialog function}
\usage{
modalDialog(
  ...,
  title = NULL,
  footer = modalButton("Dismiss"),
  size = c("m", "s", "l"),
  easyClose = FALSE,
  fade = TRUE,
  top = NULL
)
}
\description{
Mopdified modalDialog function
}
