% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{numericInput}
\alias{numericInput}
\title{Modified shiny numericInput func}
\usage{
numericInput(
  inputId,
  label,
  value,
  min = NA,
  max = NA,
  step = NA,
  width = NULL,
  tooltip = T,
  header = "?",
  popup = "Help tips",
  pos = "right"
)
}
\description{
Modified shiny numericInput func
}
