% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{prettySwitch}
\alias{prettySwitch}
\title{Modified shinyWidgets prettySwitch func}
\usage{
prettySwitch(
  inputId,
  label,
  value = FALSE,
  status = "default",
  slim = FALSE,
  fill = FALSE,
  bigger = FALSE,
  inline = FALSE,
  width = NULL,
  tooltip = T,
  header = "?",
  popup = "Help tips",
  pos = "right"
)
}
\description{
Modified shinyWidgets prettySwitch func
}
