% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{radioButtons}
\alias{radioButtons}
\title{Modified shiny radioButtons func}
\usage{
radioButtons(
  inputId,
  label,
  choices = NULL,
  selected = NULL,
  inline = FALSE,
  width = NULL,
  choiceNames = NULL,
  choiceValues = NULL,
  tooltip = T,
  header = "?",
  popup = "Help tips",
  pos = "right"
)
}
\description{
Modified shiny radioButtons func
}
