% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{selectInput}
\alias{selectInput}
\title{Modified shiny sliderInput func}
\usage{
selectInput(
  inputId,
  label,
  choices,
  selected = NULL,
  multiple = FALSE,
  selectize = TRUE,
  width = NULL,
  size = NULL,
  tooltip = T,
  title = "?",
  popup = "Help tips",
  pos = "right"
)
}
\description{
Modified shiny sliderInput func
}
