% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{sliderInput}
\alias{sliderInput}
\title{Modified shiny sliderInput func}
\usage{
sliderInput(
  inputId,
  label,
  min,
  max,
  value,
  step = NULL,
  round = FALSE,
  format = NULL,
  locale = NULL,
  ticks = TRUE,
  animate = FALSE,
  width = NULL,
  sep = ",",
  pre = NULL,
  post = NULL,
  timeFormat = NULL,
  timezone = NULL,
  dragRange = TRUE,
  tooltip = T,
  header = "?",
  popup = "Help tips",
  pos = "right"
)
}
\description{
Modified shiny sliderInput func
}
