% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_ui.R
\name{textInput}
\alias{textInput}
\title{Modified shiny textInput func}
\usage{
textInput(
  inputId,
  label,
  value = "",
  width = NULL,
  placeholder = NULL,
  tooltip = T,
  title = "?",
  popup = "Help tips",
  pos = "right"
)
}
\description{
Modified shiny textInput func
}
